Added new opmode and tested all of the limit switch code
 - Tested all limit switch code -> Works
 - Added opmode so one person can control the drive base and sample arm (not the specimen arm)
