+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/FindSlideFeedForward.java
@@ -1,12 +1,15 @@
 package org.firstinspires.ftc.teamcode.opmodes;
 
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DcMotorEx;
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
 import com.qualcomm.robotcore.util.Range;
+import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
-//@TeleOp(name="FindSlideFeedForward", group = "Real")
+@TeleOp(name="FindSlideFeedForward", group = "Real")
+@Disabled
 public class FindSlideFeedForward extends OpMode {
     DcMotorEx slideMotor;
     double powerToMotor = 0;
