+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/PIDTuning.java
@@ -1,11 +1,14 @@
 package org.firstinspires.ftc.teamcode.opmodes;
 
 import com.arcrobotics.ftclib.command.CommandScheduler;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
+import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
 import org.firstinspires.ftc.teamcode.subsystems.ArmSubsystem;
 
-//@TeleOp(name="PID Tuner", group = "Real")
+@TeleOp(name="PID Tuner", group = "Real")
+@Disabled
 public class PIDTuning extends OpMode {
     private ArmSubsystem armSubsystem;
     int armTarget;
